+++ apis/nucleus/src/__tests__/index.test.js
@@ -53,20 +53,24 @@ describe('nuked()', () => {
         let prevInstacne;
         let reactCreateElementExpectedArgs;
 
-        beforeAll(() => {
+        beforeEach(() => {
           anchorElement = document.createElement('div');
           fieldIdentifier = 'field#01';
-          options = {};
+          options = { onPopoverClose: expect.any(Function) };
           prevInstacne = document.createElement('span');
           prevInstacne.textContent = 'Previously created Instance!';
-          reactCreateElementExpectedArgs = {
-            app: undefined,
-            element: anchorElement,
-            fieldIdentifier,
-            options,
-            key: expect.any(String),
-            stateName: '$',
-          };
+          reactCreateElementExpectedArgs = [
+            expect.any(Function),
+            {
+              app: undefined,
+              element: anchorElement,
+              fieldIdentifier,
+              options,
+              key: expect.any(String),
+              stateName: '$',
+            },
+          ];
+          getListboxPopoverOptionsMock.mockReturnValue(options);
         });
 
         test('should call `root.add()` at the first time of mounting properly', () => {
@@ -78,8 +82,7 @@ describe('nuked()', () => {
           expect(getListboxPopoverOptionsMock).toHaveBeenCalledTimes(1);
           expect(getListboxPopoverOptionsMock).toHaveBeenCalledWith(options);
           expect(reactCreateElementMock).toHaveBeenCalledTimes(1);
-          expect(reactCreateElementMock.mock.lastCall[0]).toEqual(expect.any(Function));
-          expect(reactCreateElementMock.mock.lastCall[1]).toMatchObject(reactCreateElementExpectedArgs);
+          expect(reactCreateElementMock.mock.lastCall).toEqual(reactCreateElementExpectedArgs);
           expect(rootAddMock).toHaveBeenCalledTimes(1);
           expect(_nuked._popoverInstance).toEqual(createdElement);
         });
@@ -97,8 +100,7 @@ describe('nuked()', () => {
           expect(getListboxPopoverOptionsMock).toHaveBeenCalledTimes(1);
           expect(getListboxPopoverOptionsMock).toHaveBeenCalledWith(options);
           expect(reactCreateElementMock).toHaveBeenCalledTimes(1);
-          expect(reactCreateElementMock.mock.lastCall[0]).toEqual(expect.any(Function));
-          expect(reactCreateElementMock.mock.lastCall[1]).toMatchObject(reactCreateElementExpectedArgs);
+          expect(reactCreateElementMock.mock.lastCall).toEqual(reactCreateElementExpectedArgs);
           expect(rootAddMock).toHaveBeenCalledTimes(1);
           expect(_nuked._popoverInstance).toEqual(createdElement);
         });