+++ apis/nucleus/src/components/listbox/ListBox.jsx
@@ -13,6 +13,7 @@ import getListCount from './components/list-count';
 import useDataStore from './hooks/useDataStore';
 import ListBoxDisclaimer from './components/ListBoxDisclaimer';
 import ListBoxFooter from './components/ListBoxFooter';
+import getScrollIndex from './interactions/listbox-get-scroll-index';
 
 const DEFAULT_MIN_BATCH_SIZE = 100;
 
@@ -43,6 +44,7 @@ export default function ListBox({
   const local = useRef({
     queue: [],
     validPages: false,
+    dataOffset: 0,
   });
 
   const listData = useRef({
@@ -99,6 +101,23 @@ export default function ListBox({
     isSingleSelect,
   });
 
+  const { layoutOptions = {} } = layout || {};
+
+  let isRow = true;
+  if (layoutOptions.dataLayout) {
+    isRow = layoutOptions.dataLayout === 'singleColumn' ? true : layoutOptions?.layoutOrder === 'row';
+  }
+
+  const isGrid = layoutOptions?.dataLayout === 'grid';
+
+  const scrollToIndex = (index) => {
+    const gridIndex = {
+      ...(isRow ? { rowIndex: index } : { columnIndex: index }),
+    };
+    const scrollIndex = isGrid ? gridIndex : index;
+    loaderRef.current._listRef.scrollToItem(scrollIndex);
+  };
+
   const fetchData = () => {
     local.current.queue = [];
     local.current.validPages = false;
@@ -108,7 +127,8 @@ export default function ListBox({
       if (layout && layout.qSelectionInfo.qInSelections) {
         return;
       }
-      loaderRef.current._listRef.scrollToItem(0);
+      local.current.dataOffset = 0;
+      scrollToIndex(0);
     }
   };
 
@@ -133,11 +153,9 @@ export default function ListBox({
 
   useEffect(() => {
     fetchData();
-  }, [layout]);
+  }, [layout, local.current.dataOffset]);
 
   const textWidth = useTextWidth({ text: getMeasureText(layout), font: '14px Source sans pro' });
-
-  const { layoutOptions = {} } = layout || {};
 
   let minimumBatchSize = DEFAULT_MIN_BATCH_SIZE;
 
@@ -160,7 +178,30 @@ export default function ListBox({
   const { listCount } = sizes;
   setStoreValue('listCount', listCount);
 
+  const setScrollPosition = (position) => {
+    const { scrollIndex, offset, triggerRerender } = getScrollIndex({
+      position,
+      isRow,
+      sizes,
+      layout,
+      offset: local.current.dataOffset,
+    });
+    local.current.dataOffset = offset;
+    if (triggerRerender) {
+      setPages((currentPages) => [...currentPages]);
+    }
+    scrollToIndex(scrollIndex);
+  };
+
   const { textAlign } = layout?.qListObject.qDimensionInfo || {};
+
+  const [focusListItem, setFocusListItem] = useState({ first: false, last: false });
+  const getFocusState = () => ({
+    first: focusListItem.first,
+    setFirst: (first) => setFocusListItem((prevState) => ({ ...prevState, first })),
+    last: focusListItem.last,
+    setLast: (last) => setFocusListItem((prevState) => ({ ...prevState, last })),
+  });
 
   const { List, Grid } = getListBoxComponents({
     direction,
@@ -185,6 +226,8 @@ export default function ListBox({
     sizes,
     listCount,
     overflowDisclaimer: { state: overflowDisclaimer, set: showOverflowDisclaimer },
+    setScrollPosition,
+    focusListItems: getFocusState(),
   });
 
   const { columnWidth, listHeight, itemSize } = sizes || {};