+++ apis/nucleus/src/components/listbox/ListBoxInline.jsx
@@ -174,6 +174,7 @@ export default function ListBoxInline({ options = {} }) {
   const searchHeight = dense ? 27 : 40;
   const extraheight = dense ? 39 : 49;
   const minHeight = 49 + (searchVisible ? searchHeight : 0) + extraheight;
+  const wildCardSearch = layout?.wildCardSearch;
 
   const onShowSearch = () => {
     const newValue = !showSearch;
@@ -263,6 +264,7 @@ export default function ListBoxInline({ options = {} }) {
             keyboard={keyboard}
             visible={searchVisible}
             searchContainerRef={searchContainerRef}
+            wildCardSearch={wildCardSearch}
           />
         </Grid>
         <Grid item xs>